+++ src/errors/base.py
@@ -7,6 +7,7 @@ enumerators:
 """
 from dataclasses import dataclass
 from enum import Enum
+from typing import Any, Union
 
 
 @dataclass(frozen=True)
@@ -61,3 +62,8 @@ class ErrorsClassErrors(FunctionalErrorsBaseClass):
     COULD_NOT_FIND_ERROR_CODE = ErrorCode(
         code='ER_GETERROR_00001',
         description='Could not find requested error code')
+
+
+def is_error(error: Union[ErrorCode, Any]) -> bool:
+    """Method to check in object is an instance of ErrorCode."""
+    return issubclass(error.__class__, ErrorCode)